requirements.txt
openai>=1.0.0
anthropic>=0.18.0
numpy>=1.26.0
Pillow>=10.0.0
